sourcecode/requirement.txt
numpy==1.18.2
opencv-python==4.2.0.32
pefile==2019.4.18
Pillow==7.0.0
pywin32-ctypes==0.2.0
